Move user service to new folder
