+++ packages/client/src/services/user/types.ts
@@ -1,0 +1,11 @@
+import { FeedName } from '../feeds/types'
+
+interface UserFeed {
+  name: FeedName
+  isConnected: boolean
+}
+
+export interface User {
+  _id: string
+  feeds: UserFeed[]
+}
